requirements.txt
autopep8==1.5.7
et-xmlfile==1.1.0
lxml==4.6.3
openpyxl==3.0.7
Pillow==8.2.0
pycodestyle==2.7.0
python-pptx==0.6.19
six==1.16.0
toml==0.10.2
XlsxWriter==1.4.3
